booking-opt-prod/BackEnd/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx>=0.23.0
google-cloud-storage>=2.0